Determine feeder energization relation in AssignToLvFeeders.kt
